// molenapplicatie.client/src/Interfaces/Models/MolenData.ts
import { MolenImage } from './MolenImage';
import { MolenMaker } from './MolenMaker';
import { MolenTypeAssociation } from './MolenTypeAssociation';

export interface MolenData {
  id: string;
  name: string;
  ten_Brugge_Nr?: string | null;
  toelichtingNaam?: string | null;
  bouwjaar?: number | null;
  herbouwdJaar?: string | null;
  bouwjaarStart?: number | null;
  bouwjaarEinde?: number | null;
  functie?: string | null;
  doel?: string | null;
  toestand?: string | null;
  bedrijfsvaardigheid?: string | null;
  plaats?: string | null;
  adres?: string | null;
  provincie?: string | null;
  gemeente?: string | null;
  streek?: string | null;
  plaatsaanduiding?: string | null;
  opvolger?: string | null;
  voorganger?: string | null;
  opvolgerMolenId?: string | null;
  voorgangerMolenId?: string | null;
  verplaatstNaar?: string | null;
  afkomstigVan?: string | null;
  literatuur?: string | null;
  plaatsenVoorheen?: string | null;
  wiekvorm?: string | null;
  wiekVerbeteringen?: string | null;
  monument?: string | null;
  plaatsBediening?: string | null;
  bedieningKruiwerk?: string | null;
  plaatsKruiwerk?: string | null;
  kruiwerk?: string | null;
  vlucht?: string | null;
  openingstijden?: string | null;
  openVoorPubliek: boolean;
  openOpZaterdag: boolean;
  openOpZondag: boolean;
  openOpAfspraak: boolean;
  krachtbron?: string | null;
  website?: string | null;
  winkelInformatie?: string | null;
  bouwbestek?: string | null;
  bijzonderheden?: string | null;
  museuminformatie?: string | null;
  molenaar?: string | null;
  molenMakers?: MolenMaker[];
  eigendomshistorie?: string | null;
  molenerf?: string | null;
  trivia?: string | null;
  geschiedenis?: string | null;
  wetenswaardigheden?: string | null;
  wederopbouw?: string | null;
  as?: string | null;
  wieken?: string | null;
  toegangsprijzen?: string | null;
  uniekeEigenschap?: string | null;
  landschappelijkeWaarde?: string | null;
  kadastraleAanduiding?: string | null;
  canAddImages: boolean;
  eigenaar?: string | null;
  recenteWerkzaamheden?: string | null;
  rad?: string | null;
  radDiameter?: string | null;
  wateras?: string | null;
  latitude: number;
  longitude: number;
  lastUpdated?: string | Date;
  images?: MolenImage[];
  addedImages?: MolenImage[];
  molenTypeAssociations: MolenTypeAssociation[];
  hasImage: boolean;
}